requirements.txt
dnspython
pyaml
pyspf
python-whois
